microservices/customer-service/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
prometheus-flask-exporter==0.23.0
